apps/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "asyncpg>=0.30.0",
    "bitsandbytes>=0.46.0",
    "colorlog>=6.9.0",
    "crawlee[playwright]>=0.4.5",
    "dotenv>=0.9.9",
    "effdet>=0.4.1",
    "fastapi>=0.115.12",
    "funasr>=1.2.6",
    "google-cloud-vision>=3.10.1",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-experimental>=0.3.4",
    "langchain-milvus>=0.2.0",
    "langchain-xinference>=0.1.1",
    "langgraph>=0.4.8",
    "openpyxl>=3.1.5",
    "pdf2image>=1.17.0",
    "pdfminer-six==20221105",
    "pi-heif>=0.22.0",
    "pikepdf>=9.8.1",
    "protobuf>=5.29.5",
    "pydantic-settings>=2.9.1",
    "pymilvus>=2.5.10",
    "pypdf>=5.6.0",
    "pytesseract>=0.3.13",
    "python-docx>=1.1.2",
    "python-pptx>=1.0.2",
    "redis>=6.2.0",
    "sentence-transformers>=4.1.0",
    "sqlalchemy>=2.0.41",
    "starlette>=0.46.2",
    "tiktoken>=0.9.0",
    "torch>=2.7.0",
    "torchaudio>=2.7.0",
    "torchvision>=0.22.0",
    "unstructured>=0.17.2",
    "unstructured-inference>=1.0.2",
    "xinference[vllm]>=1.6.1",
]
